Rename file name to match exported class name
