+++ src/main/features/claim/routes/check-and-send.ts
@@ -5,7 +5,7 @@ import { Form } from 'forms/form'
 import { FormValidator } from 'forms/validation/formValidator'
 import { StatementOfTruth } from 'forms/models/statementOfTruth'
 import { FeesClient } from 'fees/feesClient'
-import { InterestTotal } from 'forms/models/claimInterestTotal'
+import { InterestTotal } from 'forms/models/interestTotal'
 import { InterestDateType } from 'app/common/interestDateType'
 import { claimAmountWithInterest, interestAmount } from 'utils/interestUtils'
 import { InterestType } from 'claim/form/models/interest'
